Move initialize function for orbit
